+++ src/ServiceStack.Redis/Generic/RedisTypedClient_App.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ServiceStack.Common.Utils;
 using ServiceStack.Text;
 
 namespace ServiceStack.Redis.Generic
@@ -57,7 +56,7 @@ namespace ServiceStack.Redis.Generic
 			}
 		}
 
-		public int GetRelatedEntitiesCount<TChild>(object parentId)
+		public long GetRelatedEntitiesCount<TChild>(object parentId)
 		{
 			var childRefKey = GetChildReferenceSetKey<TChild>(parentId);
 			return client.GetSetCount(childRefKey);
