+++ tests/ServiceStack.Redis.Tests/RedisTransactionTests.cs
@@ -140,9 +140,9 @@ namespace ServiceStack.Redis.Tests
 				trans.QueueCommand(r => r.AddItemToSortedSet(SortedSetKey, "sortedsetitem1"));
 				trans.QueueCommand(r => r.AddItemToSortedSet(SortedSetKey, "sortedsetitem2"));
 				trans.QueueCommand(r => r.AddItemToSortedSet(SortedSetKey, "sortedsetitem3"));
-				trans.QueueCommand(r => r.GetListCount(ListKey), intResult => collectionCounts.Add(intResult));
-				trans.QueueCommand(r => r.GetSetCount(SetKey), intResult => collectionCounts.Add(intResult));
-				trans.QueueCommand(r => r.GetSortedSetCount(SortedSetKey), intResult => collectionCounts.Add(intResult));
+                trans.QueueCommand(r => (int)r.GetListCount(ListKey), intResult => collectionCounts.Add(intResult));
+                trans.QueueCommand(r => (int)r.GetSetCount(SetKey), intResult => collectionCounts.Add(intResult));
+                trans.QueueCommand(r => (int)r.GetSortedSetCount(SortedSetKey), intResult => collectionCounts.Add(intResult));
 				trans.QueueCommand(r => r.IncrementValue(Key), intResult => incrementResults.Add(intResult));
 
 				trans.Commit();
